% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_nearest_station.R
\name{find_nearest_station}
\alias{find_nearest_station}
\title{Find nearest station}
\usage{
find_nearest_station(location = NULL, lon = NULL, lat = NULL, n = 1)
}
\arguments{
\item{n}{The number of stations to return, in order of distance}
}
\value{
A character vector of length 1
}
\description{
Find the nearest station given a string named location
}
