\name{gsod-package}
\alias{gsod-package}
\alias{gsod}
\docType{package}
\title{
gsod: global daily weather data
}
\description{
Obtain (a selection of) the Global Summary of Day (GSOD) daily weather data and interpolate it.
}
\details{
\tabular{ll}{
Package: \tab gsod\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-07-19\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}

}
\author{
Jacob van Etten

Maintainer: Jacob van Etten <jacobvanetten@yahoo.com>
}

\keyword{ package }

